Back - Just before Admin

+++ Web/app/controllers/ProfileController.php
@@ -61,7 +61,9 @@ class ProfileController extends BaseController
                 $upload           = new Upload();
                 $path             = $upload->save($this->f3->get('FILES.avatar'));
                 if (!empty($path)) {
-                    unlink($user->avatar_url);
+                    if ($user->avatar_url){
+                        unlink($user->avatar_url);
+                    }
                     $user->avatar_url = $path;
                 }
             }
